Refactored Catch proxy views into a store interface.
The goal is to make it easier to update/maintain the backend catch
proxy and also make it possible to swap for alternative backends. This
commit includes a django-native alternative backend that can be
substituted for Catch as well as an option to gather summary statistics
about annotations on a per-user basis.