Allow clones of same instance in different parts of the LMS
Use case:
Early in the course, you ask students to annotate a piece of writing. Only their annotations are shown in the sidebar
Later on, you want to show them what the instructor highlighted. So you want to show the same instance as before that has their annotations, but this time also turn on instructor annotations. This piece of code helps hide the instructor annotation in the first instance via a custom parameter in the LTI.