Working proof of concept for WYSIWYG integration.
This adds a simple interface to select an existing assignment object and
embed it on the page. The assumption is that the admin has already
created the assignment through the admin hub.

For a single content item embedded on a page, this seems to work well
enough, but it does not work for multiple content items. The problem is
that the resource_link_id is not unique in this context, but live
updates (websockets) and other parts of the app assume that it is.